infra/modules/layers/auth/requirements.txt
requests==2.33.1
python-jose==3.5.0